requirements.txt
# base
streamlit
pandas
numpy
joblib
watchdog


# 3rd-party modules
# https://streamlit.io/components
st-annotated-text

